FanControlCard: split FanFeatureSections and FanExtraRow into peer files
FanControlCard.swift was 754 lines housing four top-level types. Splits:

- FanFeatureSections (the native List sections variant for use under
  iOS-style settings screens) → Shared/FanControl/FanFeatureSections.swift.
  Mirrors the same pattern as Shared/LightControl/LightFeatureSections.swift.
- FanExtraRow (binary/enum/numeric/text row used inside the card's extras
  list) → Shared/FanControl/FanExtraRow.swift, becoming a top-level type so
  it can be split out without exposing more API than the card needs (it was
  already file-private; now file-private to its own file).

DisclosureRow stays inside FanControlCard.swift — it's a private helper
used only there, so it's tightly-coupled per the new code-structure rule.

Both new files are added to the ShellbeeWidgetsExtension membershipExceptions
list alphabetically, since they reference main-app types (Expose, JSONValue,
FeatureCatalog) that don't ship in the widget target.

Brings FanControlCard.swift from 754 → 538 lines (still over the 400 soft
cap but below the 600 hard cap, and the remainder is one cohesive type).

Refs #33
